backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
numpy
pydantic
python-dotenv
dashscope
agentscope
